Added unit tests for FixedRateCoupon, FloatingRateCoupon, and Swap

+++ QFF/UnitTest/DayCounterStub.h
@@ -4,7 +4,7 @@
 
 class DayCounterStub : public IDayCounter {
 public:
-	MOCK_CONST_METHOD2(countDayBetween, int(date, date));
-	MOCK_CONST_METHOD2(calculateYearFraction, int(date, date));
+	MOCK_CONST_METHOD2(countDayBetween, double(date, date));
+	MOCK_CONST_METHOD2(calculateYearFraction, double(date, date));
 };
